Add ability to pause/relaunch app to import db
