+++ helpers/app-states.js
@@ -1,0 +1,6 @@
+'use strict'
+
+module.exports = {
+  isMainWinMaximized: false,
+  loadURL: null
+}
