+++ helpers/index.js
@@ -14,6 +14,20 @@ const {
 const windowStateKeeper = require('./window-state-keeper')
 const exportDB = require('./export-db')
 const importDB = require('./import-db')
+const windows = require('./windows')
+const ipcs = require('./ipcs')
+const runServer = require('./run-server')
+const createMenu = require('./create-menu')
+const pauseApp = require('./pause-app')
+const relaunch = require('./relaunch')
+const appStates = require('./app-states')
+const {
+  createMainWindow,
+  createLoadingWindow,
+  createErrorWindow
+} = require('./window-creators')
+const initializeApp = require('./initialize-app')
+const makeSingleInstance = require('./make-single-instance')
 
 module.exports = {
   bootTwoGrapes,
@@ -23,5 +37,17 @@ module.exports = {
   checkAndChangeAccess,
   windowStateKeeper,
   exportDB,
-  importDB
+  importDB,
+  windows,
+  ipcs,
+  runServer,
+  createMenu,
+  pauseApp,
+  relaunch,
+  appStates,
+  createMainWindow,
+  createLoadingWindow,
+  createErrorWindow,
+  initializeApp,
+  makeSingleInstance
 }
