+++ helpers/ipcs.js
@@ -1,0 +1,5 @@
+'use strict'
+
+module.exports = {
+  serverIpc: null
+}
