+++ helpers/layouts/app-init.html
@@ -4,7 +4,7 @@
 <head>
   <meta charset="utf-8">
   <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-  <link rel="shortcut icon" href="../bfx-report-ui/build/favicon.ico">
+  <link rel="shortcut icon" href="../../bfx-report-ui/build/favicon.ico">
   <title>Bitfinex Reports</title>
   <style>
     html {
@@ -188,7 +188,7 @@
     <div></div>
   </div>
   <div class="logo">
-    <img class="logo-img" src="../build/icons/512.png" alt="logo">
+    <img class="logo-img" src="../../build/icons/512.png" alt="logo">
   </div>
 </body>
