+++ helpers/layouts/express-port-required.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-    <link rel="shortcut icon" href="../bfx-report-ui/build/favicon.ico">
+    <link rel="shortcut icon" href="../../bfx-report-ui/build/favicon.ico">
     <link href="https://fonts.googleapis.com/css?family=Roboto" rel="stylesheet">
     <title>Bitfinex Reports</title>
     <style>
