+++ helpers/windows.js
@@ -1,0 +1,7 @@
+'use strict'
+
+module.exports = {
+  mainWindow: null,
+  loadingWindow: null,
+  errorWindow: null
+}
